+++ src/main/java/com/joao/sistema_estoque/service/ClienteService.java
@@ -2,17 +2,19 @@ package com.joao.sistema_estoque.service;
 
 import com.joao.sistema_estoque.model.Cliente;
 import com.joao.sistema_estoque.repository.ClienteRepository;
+import com.joao.sistema_estoque.repository.VendaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import com.joao.sistema_estoque.exception.ResourceNotFoundException;
-
+import com.joao.sistema_estoque.exception.BusinessException;
 
 @Service
 @RequiredArgsConstructor
 public class ClienteService {
 
     private final ClienteRepository repository;
+    private final VendaRepository vendaRepository;
 
     public List<Cliente> listarTodos() {
         return repository.findAll();
@@ -39,6 +41,9 @@ public class ClienteService {
 
     public void deletar(Long id) {
         buscarPorId(id);
+        if (!vendaRepository.findByCliente_Id(id).isEmpty()) {
+            throw new BusinessException("Cliente está vinculado a uma ou mais vendas e não pode ser excluído.");
+        }
         repository.deleteById(id);
     }
 }
